+++ CHANGELOG.md
@@ -1,13 +1,11 @@
 Change Log
 ==========
 
-## Version 2.2.1
-- Resolve gradle check errors
-
 ## Version 2.2.0
 - Requesting token now uses PKCE when:
     - There is an installed Spotify app that supports PKCE.
     - When web fallback is used.
+- Token requests may return a refresh token if PKCE was used.
 
 ## Version 2.1.2
 * Propagate tracking parameters when opening the native login flow